requirements.txt
# Core AI 및 LLM 관련
openai>=1.6.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.10
langgraph>=0.0.15
pydantic>=2.5.2

# 데이터 처리 및 시각화
numpy>=1.24.0
pandas>=2.1.3
plotly>=5.18.0

# 웹 인터페이스
streamlit>=1.28.2
streamlit-chat>=0.1.1

# 문서 처리
PyMuPDF>=1.23.7
pytesseract>=0.3.10
Pillow>=10.1.0

# 딥러닝 및 임베딩
torch>=2.1.1
transformers>=4.35.2
sentence-transformers>=2.2.2

# 환경 설정 및 유틸리티
python-dotenv>=1.0.0
tqdm>=4.66.1
requests>=2.31.0
xmltodict>=0.13.0

# PDF 및 이미지 처리 의존성
pdf2image>=1.16.3
poppler-utils; sys_platform != "win32"
tesseract-ocr; sys_platform != "win32"